bugfix alertRuleStore & test code
